requirements.txt
beautifulsoup4==4.6.0
boto3==1.7.11
botocore==1.10.11
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
futures==3.2.0
gspread==3.0.0
httplib2==0.11.3
idna==2.6
jmespath==0.9.3
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.2
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
